Restructure project layout to allow mixing of pure Python and compiled module
- c++ library code in /lib
- pybind11 c++/python binding code in /src
- pure python code in /src/pybind11_numpy_example
  - add example pure python function `pure_python_list()`
- replace `-` with `_` in all filenames and paths
- update plots
