+++ CMakeLists.txt
@@ -3,14 +3,14 @@ cmake_minimum_required(VERSION 3.15..3.26)
 # Set a name and a version number for your project:
 project(
   pybind11-numpy-example
-  VERSION 0.0.7
+  VERSION 0.0.9
   LANGUAGES CXX)
 
 # Initialize some default paths
 include(GNUInstallDirs)
 
 # Define the minimum C++ standard that is required
-set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
@@ -19,15 +19,15 @@ set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 option(BUILD_PYTHON "Enable building of Python bindings" ON)
 option(BUILD_DOCS "Enable building of documentation" ON)
 
-# Build the library
-add_subdirectory(src)
+# Build the c++ library
+add_subdirectory(lib)
 
-# Build the tests
+# Build the c++ tests
 include(CTest)
 if(BUILD_TESTING)
   add_subdirectory(ext/Catch2)
   include(./ext/Catch2/extras/Catch.cmake)
-  add_subdirectory(tests)
+  add_subdirectory(tests/cpp)
 endif()
 
 # Build the documentation
@@ -37,26 +37,26 @@ endif()
 
 # Build the python bindings
 if(BUILD_PYTHON)
-  add_subdirectory(python)
+  add_subdirectory(src)
 endif()
 
 # Add an alias target for use if this project is included as a subproject in
 # another project
-add_library(pybind11-numpy-example::pybind11-numpy-example ALIAS
-            pybind11-numpy-example)
+add_library(pybind11_numpy_example::pybind11_numpy_example ALIAS
+            pybind11_numpy_example)
 
 # Install targets and configuration
 install(
-  TARGETS pybind11-numpy-example
-  EXPORT pybind11-numpy-example-config
+  TARGETS pybind11_numpy_example
+  EXPORT pybind11_numpy_example_config
   RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
   LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
   ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR})
 
 install(
-  EXPORT pybind11-numpy-example-config
-  NAMESPACE pybind11-numpy-example::
-  DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/pybind11-numpy-example)
+  EXPORT pybind11_numpy_example_config
+  NAMESPACE pybind11_numpy_example::
+  DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/pybind11_numpy_example)
 
 install(DIRECTORY ${CMAKE_CURRENT_LIST_DIR}/include/
         DESTINATION ${CMAKE_INSTALL_INCLUDEDIR})
