+++ scripts/mem_array.txt
@@ -1,15 +1,22 @@
 #n memory (kb)
-1000 18896
-10000 18864
-100000 19560
-1000000 23260
-10000000 58380
-50000000 214536
-100000000 410024
-200000000 800452
-300000000 1191012
-400000000 1581412
-1000000000 3925392
-2000000000 7831684
-3000000000 11737892
-4000000000 15644044
+1000 20648
+10000 20664
+100000 20840
+1000000 24520
+10000000 59672
+50000000 215920
+100000000 411412
+200000000 801848
+300000000 1192548
+400000000 1583112
+600000000 2363992
+800000000 3145808
+1000000000 3926680
+1200000000 4708296
+2000000000 7832800
+3000000000 11739192
+4000000000 15645488
+6000000000 23458288
+8000000000 31270592
+10000000000 39083108
+12000000000 46895616
