+++ scripts/mem_array_nocopy.txt
@@ -1,19 +1,27 @@
 #n memory (kb)
-1000 18808
-10000 18864
-100000 18936
-1000000 20912
-10000000 38284
-50000000 116396
-100000000 214164
-200000000 409508
-300000000 604768
-400000000 800084
-1000000000 1971984
-2000000000 3925016
-3000000000 5878236
-4000000000 7831396
-5000000000 9784604
-6000000000 11737484
-7000000000 13690668
-8000000000 15643960
+1000 20656
+10000 20668
+100000 20844
+1000000 22596
+10000000 39936
+50000000 118080
+100000000 215964
+200000000 411264
+300000000 606336
+400000000 801716
+600000000 1192512
+800000000 1583148
+1000000000 1973368
+1200000000 2364404
+2000000000 3926908
+3000000000 5880000
+4000000000 7832964
+6000000000 11739264
+8000000000 15645624
+10000000000 19551908
+12000000000 23457968
+14000000000 27364404
+16000000000 31270648
+18000000000 35176512
+20000000000 39083136
+24000000000 46895548
