+++ scripts/mem_list.txt
@@ -1,11 +1,15 @@
 #n memory (kb)
-1000 0
-10000 444
-100000 4164
-1000000 41356
-10000000 414140
-50000000 2069920
-100000000 4139380
-200000000 8278588
-300000000 12418828
-400000000 16557252
+1000 192
+10000 384
+100000 4224
+1000000 41088
+10000000 410304
+50000000 2050752
+100000000 4101504
+200000000 8203200
+300000000 12304704
+400000000 16405824
+600000000 24609216
+800000000 32811840
+1000000000 41014464
+1200000000 49218432
