+++ scripts/time_array.txt
@@ -1,15 +1,22 @@
 #n time (seconds)
-1000 1.0294279672216436e-05
-10000 9.17243686327655e-05
-100000 0.0008979495345340525
-1000000 0.008858304726097478
-10000000 0.053286767499230336
-50000000 0.13255561800906435
-100000000 0.17452786200738046
-200000000 0.25977244100067765
-300000000 0.3446448600006988
-400000000 0.4255073300009826
-1000000000 0.9024126849981258
-2000000000 1.6776944080047542
-3000000000 2.439360503994976
-4000000000 3.1745616889966186
+1000 8.885241385522114e-06
+10000 8.064530965231068e-05
+100000 0.0007929159005605938
+1000000 0.009228698179041121
+10000000 0.05545463500311598
+50000000 0.16033962194342166
+100000000 0.23489250801503658
+200000000 0.387445722008124
+300000000 0.5446825119433925
+400000000 0.6931000800104812
+600000000 1.0074213709449396
+800000000 1.3038605999900028
+1000000000 1.6112821800634265
+1200000000 1.8958396930247545
+2000000000 3.080139480996877
+3000000000 4.569697203929536
+4000000000 6.015317940968089
+6000000000 8.836932464968413
+8000000000 11.728594742016867
+10000000000 14.558276921976358
+12000000000 17.50521888397634
