+++ scripts/time_array_nocopy.txt
@@ -1,19 +1,27 @@
 #n time (seconds)
-1000 9.999668932844945e-06
-10000 9.167261039231114e-05
-100000 0.0008991213465261614
-1000000 0.008335511999162422
-10000000 0.050443788000848144
-50000000 0.12099327699979767
-100000000 0.15236243599792942
-200000000 0.21317349300079513
-300000000 0.27599514700705186
-400000000 0.33998234900354873
-1000000000 0.6818497240019497
-2000000000 1.2458708429912804
-3000000000 1.7789631179912249
-4000000000 2.310697248991346
-5000000000 2.809482886994374
-6000000000 3.338349138997728
-7000000000 3.847873509992496
-8000000000 4.3607585489953635
+1000 8.883211479922996e-06
+10000 7.990702798778003e-05
+100000 0.0007843043065012092
+1000000 0.007740272727625614
+10000000 0.050859106006100774
+50000000 0.1441949950531125
+100000000 0.239397105993703
+200000000 0.3419582910137251
+300000000 0.48297904699575156
+400000000 0.6020817440003157
+600000000 0.8661073229741305
+800000000 1.1198496220167726
+1000000000 1.3745229849591851
+1200000000 1.6183152759913355
+2000000000 2.622669712989591
+3000000000 3.8534785190131515
+4000000000 5.067567399004474
+6000000000 7.537025678087957
+8000000000 9.918707602075301
+10000000000 12.325743645080365
+12000000000 14.685017141047865
+14000000000 17.10181719996035
+16000000000 19.59815236995928
+18000000000 21.906724045053124
+20000000000 24.42910428403411
+24000000000 29.188139538047835
