+++ scripts/time_list.txt
@@ -1,11 +1,15 @@
 #n time (seconds)
-1000 1.6527705828344668e-05
-10000 0.00021758406994293084
-100000 0.0027138690693571353
-1000000 0.03301038918222978
-10000000 0.3295107399972039
-50000000 1.6292891499906546
-100000000 3.3216604850022122
-200000000 6.464595986995846
-300000000 9.88627804099815
-400000000 12.484707872994477
+1000 1.1281892404451172e-05
+10000 0.00012381860035491158
+100000 0.0021609368014822504
+1000000 0.028091967190531166
+10000000 0.3382009760243818
+50000000 1.6539601690601557
+100000000 3.3066364750266075
+200000000 6.4062930109212175
+300000000 9.456814253004268
+400000000 12.655741214985028
+600000000 18.71708781796042
+800000000 24.94151852594223
+1000000000 31.0570098310709
+1200000000 37.80632794194389
