+++ src/CMakeLists.txt
@@ -1,5 +1,4 @@
-add_library(pybind11-numpy-example pybind11-numpy-example.cpp)
-target_include_directories(
-  pybind11-numpy-example
-  PUBLIC $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/include/>
-         $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>)
+find_package(pybind11 CONFIG REQUIRED)
+pybind11_add_module(_pybind11_numpy_example pybind11_numpy_example_python.cpp)
+target_link_libraries(_pybind11_numpy_example PUBLIC pybind11_numpy_example)
+install(TARGETS _pybind11_numpy_example DESTINATION pybind11_numpy_example)
