--- src/pybind11-numpy-example.cpp
@@ -1,1 +1,0 @@
-#include "pybind11-numpy-example/pybind11-numpy-example.hpp"
